clightning-plugins/gracefulshutdown/requirements.txt
pyln-client ~= 0.9.2
